+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] - 2018-05-05
+
+### Added
+
+* Functions that calculate water properties (for aqueous reactions).
 
 ## 0.0.1 - 2018-05-04